Refine topics to clarify scenarios and add keywords
This re-writes some of the questions in the list of topics that the workshop
will cover, with a view to introducing a few additional keywords and clarifying
the different scenarios that media companies are usually interested in:
broadcast/live, catch-up, VOD, and offline distribution mechanisms. Plus the
use of edge cache servers for OTT delivery.